+++ framework/CHANGELOG.md
@@ -46,6 +46,7 @@ Yii Framework 2 Change Log
 - Enh #13695: `\yii\web\Response::setStatusCode()` method now returns the Response object itself (kyle-mccarthy)
 - Enh #13698: `yii\grid\DataColumn` filter is automatically generated as dropdown list in case of `format` set to `boolean` (bizley)
 - Enh #13254: Core validators no longer require Yii::$app to be set (sammousa)
+- Bug #4408: Add support for unicode word characters and `+` character in attribute names (sammousa, kmindi)
 - Bug #10372: Fixed console controller including complex typed arguments in help (sammousa)
 
 2.0.11.2 February 08, 2017